requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
arrow==1.3.0
asgiref==3.8.1
attrs==25.3.0
backoff==2.2.1
blis==1.3.0
boto3==1.38.27
botocore==1.38.27
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.1
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
cymem==2.0.11
dataclasses-json==0.6.7
deepeval==3.0.2
Deprecated==1.2.18
diff-match-patch==20230430
distro==1.9.0
dj-database-url==1.2.0
dj-rest-auth==7.0.1
Django==5.1
django-allauth==65.7.0
django-cors-headers==4.7.0
djangorestframework==3.15.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
email_validator==2.2.0
en_core_web_md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl#sha256=5e6329fe3fecedb1d1a02c3ea2172ee0fede6cea6e4aefb6a02d832dba78a310
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
execnet==2.1.1
Faker==25.9.2
filelock==3.18.0
flatbuffers==25.2.10
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2025.5.1
google-auth==2.40.2
google-genai==1.18.0
googleapis-common-protos==1.70.0
griffe==0.36.9
grpcio==1.71.0
guardrails-ai==0.6.6
guardrails-api-client==0.4.0a1
guardrails_hub_types==0.0.4
gunicorn==21.2.0
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.3
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.1.0
isoduration==20.11.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.63
langchain-openai==0.3.18
langchain-text-splitters==0.3.8
langcodes==3.5.0
langsmith==0.3.43
language_data==1.3.0
litellm==1.72.0
lxml==4.9.4
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.4.4
murmurhash==1.0.13
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
newrelic==10.12.0
numpy==2.2.6
ollama==0.5.1
onnx==1.18.0
onnxruntime==1.22.0
openai==1.82.1
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-exporter-otlp-proto-http==1.33.1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
optimum==1.25.3
orjson==3.10.18
packaging==24.2
pillow==11.2.1
pluggy==1.6.0
portalocker==3.1.1
posthog==3.25.0
preshed==3.0.10
propcache==0.3.1
protobuf==5.29.5
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydash==7.0.7
Pygments==2.19.1
PyJWT==2.9.0
pypdf==5.5.0
pytest==8.3.5
pytest-asyncio==1.0.0
pytest-repeat==0.9.4
pytest-rerunfailures==12.0
pytest-xdist==3.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2023.3
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rest-framework-simplejwt==0.0.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.25.1
rsa==4.9.1
rstr==3.2.2
s3transfer==0.13.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
semver==3.0.4
sentence-transformers==4.1.0
sentry-sdk==2.29.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
sqlparse==0.5.3
srsly==2.5.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.0.0
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.9.0
timm==1.0.15
tokenizers==0.21.1
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.51.3
typer==0.15.4
types-python-dateutil==2.9.0.20250516
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
uri-template==1.3.0
urllib3==2.0.7
wasabi==1.1.3
weasel==0.4.1
webcolors==24.11.1
websockets==15.0.1
wheel==0.45.1
whitenoise==6.5.0
wrapt==1.17.2
yarl==1.20.0
zipp==3.22.0
zstandard==0.23.0
